Moved GroupingEventVM to ViewModel.
