+++ ModelLibrary/Event.cs
@@ -12,8 +12,9 @@ namespace ModelLibrary
         public string Abstract { get; set; }
         public int EventID { get; set; }
         public int RoomNr { get; set; }
+        public string ImagePath { get; set; }
 
-        public Event(DateTime date, TimeSpan duration, float rating, string @abstract, int eventId, int roomNr)
+        public Event(DateTime date, TimeSpan duration, float rating, string @abstract, int eventId, int roomNr, string imagePath)
         {
             Date = date;
             Duration = duration;
@@ -21,6 +22,7 @@ namespace ModelLibrary
             Abstract = @abstract;
             EventID = eventId;
             RoomNr = roomNr;
+            ImagePath = imagePath;
         }
 
         public Event()
