+++ SikonConferenceSystem/ViewModel/GroupingEventVM.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using ModelLibrary;
 
-namespace SikonConferenceSystem.View
+namespace SikonConferenceSystem.ViewModel
 {
+    /// <summary>
+    /// Demo Class. Susceptible to change.
+    /// </summary>
     public class GroupingEventVM
     {
         public ObservableCollection<HourGroup> HourGroups { get; set; }
@@ -20,24 +20,24 @@ namespace SikonConferenceSystem.View
 
             List<EventAdapter> allEvents = new List<EventAdapter>();
             
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(11), TimeSpan.FromHours(1), 5, "The first event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(11), TimeSpan.FromHours(2), 5, "The second event", 0, 0),
-                "#33EA1616", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(12), TimeSpan.FromHours(1), 5, "The Third event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(12), TimeSpan.FromHours(1), 5, "The Fourth event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13), TimeSpan.FromHours(2), 5, "The Fifth event", 0, 0),
-                "#33EA1616", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13), TimeSpan.FromHours(1), 5, "The Sixth event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(14), TimeSpan.FromHours(1), 5, "The Seventh event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(14), TimeSpan.FromHours(1), 5, "The Eight event", 0, 0),
-                "#33DEF010", ""));
-            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13.5), TimeSpan.FromHours(1), 5, "The Test event", 0, 0),
-                "#33DEF010", ""));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(11), TimeSpan.FromHours(1), 5, "The first event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(11), TimeSpan.FromHours(2), 5, "The second event", 0, 0, ""),
+                "#33EA1616"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(12), TimeSpan.FromHours(1), 5, "The Third event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(12), TimeSpan.FromHours(1), 5, "The Fourth event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13), TimeSpan.FromHours(2), 5, "The Fifth event", 0, 0, ""),
+                "#33EA1616"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13), TimeSpan.FromHours(1), 5, "The Sixth event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(14), TimeSpan.FromHours(1), 5, "The Seventh event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(14), TimeSpan.FromHours(1), 5, "The Eight event", 0, 0, ""),
+                "#33DEF010"));
+            allEvents.Add(new EventAdapter(new Event(DateTime.Today.AddHours(13.5), TimeSpan.FromHours(1), 5, "The Test event", 0, 0, ""),
+                "#33DEF010"));
 
             SetupHourGroups();
 
@@ -60,7 +60,7 @@ namespace SikonConferenceSystem.View
                 {
                     if (AddCondition(@event, hourGroup))
                     {
-                        hourGroup.EventGroup.Add(@event);
+                        hourGroup.Events.Add(@event);
                     }
                 }
             }
@@ -78,11 +78,11 @@ namespace SikonConferenceSystem.View
 
     public class HourGroup : ObservableCollection<EventAdapter>
     {
-        public ObservableCollection<EventAdapter> EventGroup { get; set; }
+        public ObservableCollection<EventAdapter> Events { get; set; }
 
         public HourGroup(DateTime time)
         {
-            EventGroup = new ObservableCollection<EventAdapter>();
+            Events = new ObservableCollection<EventAdapter>();
             TimeAsDateTime = time;
             Time = $"{time.ToShortTimeString()} : {time.AddHours(1).ToShortTimeString()}";
         }
@@ -96,13 +96,11 @@ namespace SikonConferenceSystem.View
     {
         public Event Event { get; set; }
         public string Color { get; set; }
-        public string ImagePath { get; set; }
 
-        public EventAdapter(Event @event, string color, string imagePath)
+        public EventAdapter(Event @event, string color)
         {
             Event = @event;
             Color = color;
-            ImagePath = imagePath;
         }
     }
 }
